Merged the functionality of the cpdftocps filter into pdftops
There is a "cpdftocps" filter to turn already page-managed PDF
(application/vnd.cups-pdf) into page-managed and
PPD-option-setting-code-loaded PostScript
(application/vnd.cups-postscript). This filter is a shell script
calling Poppler's pdftops utility and and CUPS' pstops filter without
applying page management options. The filter is downloadable here:

http://www.openprinting.org/download/printing/pdf-printing/cpdftocps
http://www.openprinting.org/download/printing/pdf-printing/cpdftocps.convs

In the PDF printing workflow this filter is meant to be the
"PostScript printer driver" as it generates PostScript output for
PostScript printers and inserts the option code of the
manufacturer-supplied PPD files (which implicitly require
application/vnd.cups-postscript as input) into the PostScript.

Comparing this filter with CUPS' "pdftops" filter (which turns
incoming PDF print jobs into PostScript and then applies page
management and PPD options to that PostScript) one sees that the
functionality is nearly identical.

Therefore we merge these two filters into one with this commit. The
filter recognizes on the input data whether it has to behave like the
"pdftops" or like the "cpdftocps" filter. Also the MIME conversion
rule for the "cpdftocps" filter is added.