Added first initial version of the bannertopdf filter
The new bannertopdf filter is a replacement for the bannertops filter
and completes the filter set needed for the PDF-based printing
workflow.

Feature-wise, the bannertopdf filter supports PDF template pages,
one-page PDF files which make up the background for test pages and
banners. This allows to customize banners and test pages for branding
and for adding vector graphics and color management tests to the test
objects on a test page.

Thanks to Lars Uebernickel from Canonical to create this filter.

In addition, the build system, the MIME conversion rules, and the
Dependencies file got updated.