Re-worked the cost factors for the filters
Especially avoid pstotiff|imagetopdf instead of pstopdf/pdftopdf for
PostScript->PDF conversion.

Many PDF-based filters have now the same cost factor as their former
PostScript-based equivalent of CUPS but reduced by one so that if
cups-filters is used with an old CUPS version the PDF-based filter of
cups-filters is preferred.