utils/cups-browsed.c: Lock before all 'examine_discovered_printer_record()'
The current code doesn't use rwlock in found_cups_printer() and
cupsdUpdateLDAPBrowse(), which causes deadlock when discovering queues
via BrowsePoll or LDAP.
It happens because the code in 'examine_discovered_printer_record()'
expects the rwlock to be locked already, so it unlocks the lock, calls
'get_local_queue_name()' and locks again.

The commit contains change '&(lock)' -> '&lock' too, it represents the
same thing, so we should use the same code for it.