libcupsfilters: Let cupsRasterPrepareHeader() create complete header
The cupsRasterPrepareHeader() function creates a CUPS/PWG Raster
header for Raster output based on the printer and job properties
supplied to the calling filter functions.

This commit fixes the header produced by this function to be complete
and consistent so that it can be used for actual Raster output. It
adds the following missing pieces:

- Find correct dimensions and margins for the page size requested by
  the job.

- After having found page dimensions, margins, color space, color
  depth, and resolution, set all fields of the header which depend on
  this to the correct values, to assure that the resulting header is
  consistent.

Also did some fixes in cupsRasterParseIPPOptions()

- Create the empty header with all-zero before starting, to have
  everything well-defined.

- Fixed setting for requested orientation, to correctly handle 0 as
  input or wrong input values.