libppd: When running driver/PPD generator executables, capture stderr
When generating PPD lists for PPD collections or extracting a PPD file
from a collection, CUPS driver/PPD generator executables (CUPS has
them in /usr/lib/cups/driver) can be called. The stderr output
(fortunately rare) of the driver/PPD generator executables did not get
captured and went directly to stderr (to syslog when the caller runs
as a system daemon).

This commit fixes it and re-directs the stderr output into the log
function, so that nothing is leaking into stderr any more.

In addition, more logging is added to the PipeCommand() and
ClosePipeCommand() functions.