imagetoraster, imagetopdf: Fixed "print-scaling=none"
In the imageto...()s fixed the print-scaling=none (crop-to-fit) mode,
which prints the image in its original size (based on PPI resolution
stored in the image file or on 200 dpi resolution, which is the
standard resolution for shipping labels). Also use crop-to-fit always
when requested, do not fall back to fit-to-page when the image is
significantly larger or smaller than the page.

(manually backported from commit cb2737b5683adfbe5cc25bada323f43cc1ca2329)