libcupsfilters: Updated comment about human-readable strings in PPD Generator
There is a longer comment about how to obtain and handle
human-readable strings and translations in the PPD file generator for
driverless printing, right before the source code of the PPD file
generator.

This comment tells that there are no human-readable strings for the
manual color space and depth choices in the ColorModel option wheras
IPP has only simple bi-level/grayscale/color choices and the exact
color space and depth the filters are supposed to select automatically
according to job and printer.

This part of the comment got already obsolete a few days later when I
made the filters auto-selecting color space and depth via the newly
introduced cupsRasterPrepareHeader() and cupsRasterSetColorSpace()
functions.

Now I discovered this forgotten comment and removed the obsolete part.