libppd: In ppdLoadAttributes() complete custom page size support
If the user has sent a job on a custom page size, these custom size is
correctly set as default size in the printer IPP attributes derived
from the PPD file, as media-col, with dimensions, margins, and PWG
specifying the size in the units originally used for it.

Also added the custom page size range keywords to the
"media-supported" printer IPP attribute.