libcupsfilters: Added RGBW support to cfFilterPCLmToRaster()
Requesting RGBW as destination color space (to test, use pclmtoraster
CUPS filter wrapper and PPD for HP inkjet with hpcups driver of HPLIP)
leads to wrong color output on some input files and even to a crash on
others.

This is cause by missing support for RGBW and this, instead RGB being
output into the destination file with RGBW declared as the color space
in its header.

This commit adds explicit support for RGBW output and this way solves
the problem.