libppd: Code clean-up for code of foomatic-rip
Cleaned up the code following the coding style rules in the
DEVELOPING.md file of the CUPS source code.

Comments are re-formatted to use "// ..." instead of "/* ... */", like
in PAPPL, so C and C++ files get the same comment style.

The indent width is set to 2 spaces instead of 4, to match the other
files.

This way we get rid of the mix of many different coding styles which
came together from the many code contributions received during more
than a decade, even before the start of the cups-filters project.