In auto-generated PPDs do not set RGB default on mono printers
When a PPD for a driverless printer is generated by the
_ppdCreateFromIPP2() function and the get-printer-attributes IPP
response gives "print-color-mode-default=auto" the PPD's default
setting for "ColorModel" is always "RGB", even on monochrome printers,
which makes printing fail on most devices.

Now we ignore the "print-color-mode-default" if set to "auto" and
proceed as if no default was given, finding the default by selecting
the most desirable of the existing "ColorModel" choices.

Discovered via

https://github.com/OpenPrinting/cups/issues/614

The bug was originally reported against ipp-usb as the reporter's
printer was connected via IPP-over-USB, but actually the PPD
generators are at fault, and here not only the one in CUPS but also
the one in cups-filters 1.x and libppd 2.x.