PPD file generator: Support more than 2 Apple Raster resolutions
Backport of OpenPrinting/libppd@0b2ca66 to cups-filters 1.x.

Which does the following for cupsPrintQuality:
Draft Quality: Lowest resolution
High Quality: Highest resolution
Normal Quality: If there is printer-default-resolution and this resolution
is in the list, use it.  If not, use the "middle" one, rounding down if the
number of available resolutions is even.

Backport was simplified further by merging two while loops into one and
consequently removing redundant variables, it was also modified to replace
a 128KB, 32767 element array with a 16 element array used for temporary
storage of the supported Apple raster resolutions.