filter: add a raster to ESC/POS filter
ESC/POS is a de facto standard for POS (Point Of Sale) terminal
printers (eg, thermal receipt printers).

This patch adds a very bare-bones implementation to print raster data
on printers that implement this command set. It also includes support
for cutting paper either at the end of a page or at the end of a job.

Signed-off-by: Sergio Gómez <sergio.g.delreal@gmail.com>