Introduce foomatic-hash and reject unauthorized values in foomatic-rip
The PR provides a way for users to have control over what values are
allowed for foomatic-rip related PPD keywords FoomaticRIPCommandLine,
FoomaticRIPCommandLinePDF, and FoomaticRIPOptionSetting. Since the
values can be later used when constructing shell command, the filter
foomatic-rip was a target of several exploits (caused by issue at
different place in CUPS or in different project of printing stack) to
do arbitrary code execution when the filter is started.

By default the filter is run by user lp, so the issue is mitigated, but
this PR gives admin complete control over what can be run in
foomatic-rip and reject anything injected into system via different
ways.

First, the new tool - foomatic-hash - can be called on a PPD file or
directory with drivers/PPDs, with scan output and file with hexadecimal
representation of hashed values. Once the scan output is reviewed by
admin, admin can decide to put the resulting hashes into
/etc/foomatic/hashes.d and allow them for the filter.